pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devwatcher"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "watchdog>=4.0",
    "gitpython>=3.1",
    "pywin32>=306; sys_platform == 'win32'",
    "anthropic>=0.40",
]

[project.scripts]
devwatcher = "devwatcher.cli:app"

[project.optional-dependencies]
gemini = ["google-genai>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
